Change phabricator to talk to the pygmentize server directly.
Before, we were telling it to talk to a shell script that called out
to `curl`, so we wouldn't have to change the php code.  But that is a
lot less efficient: we have to exec the shell and then exec curl (and
also exec a `sed` command to get the input in the right form).  It's a
lot simpler to just talk to the http server directly!  And not that
hard.

This is a port of the same change made in libphutil, in D43073 and
some follow-up bugfix audits.  I am porting it here to follow
upstream, which has merged all of libphutil into phabricator or
arcanist (depending on functionality).